refactor(webapp): use ProducersTable component
