+++ webapp/src/routes/NonCompliantBPs/index.js
@@ -5,12 +5,23 @@ import LinearProgress from '@mui/material/LinearProgress'
 
 import useNonCompliantState from '../../hooks/customHooks/useNonCompliantState'
 import NoResults from '../../components/NoResults'
-import UndiscoverableBPsTable from '../../components/ProducersList/UndiscoverableBPsTable'
+import ProducersTable from '../../components/ProducersList/ProducersTable'
+import NonCompliantCard from '../../components/NonCompliantCard'
 
 import RewardsStats from './RewardsStats'
 
 const NonCompliantBPs = () => {
   const [{ items, stats, loading }] = useNonCompliantState()
+  const columnsNames = [
+    'rank',
+    'producerName',
+    'website',
+    'bpJson',
+    'votes',
+    'lastClaimTime',
+    'dailyRewards',
+    'yearlyRewards',
+  ]
 
   return (
     <>
@@ -22,11 +33,16 @@ const NonCompliantBPs = () => {
             <>
               <RewardsStats stats={stats} />
               <Card>
-                 <UndiscoverableBPsTable producers={items} tokenPrice={stats.tokenPrice}/>
+                <ProducersTable
+                  columnsNames={columnsNames}
+                  producers={items}
+                  RowComponent={NonCompliantCard}
+                  RowProps={{ tokenPrice: stats.tokenPrice || NaN }}
+                />
               </Card>
             </>
           ) : (
-            <NoResults translationScope={'undiscoverableBPsRoute'}/>
+            <NoResults translationScope={'undiscoverableBPsRoute'} />
           )}
         </>
       )}
